Try to fix 0.99999999 times

+++ src/base/helper_kz.cpp
@@ -3,31 +3,31 @@
 #include "system.h"
 #include <cmath>
 
-const char * str_format_time_kz(float Time)
+const char * str_format_time_kz(double Time)
 {
     static char stringvar[256];
-	float tempvar, floattime;
+	double tempvar, floattime;
 	floattime = ((int)Time) % 60 + std::modf(Time, &tempvar);
 	if(Time >= 60.f) //+KZ minutes
 	{
 		if(Time / 60.f >= 60) //+KZ hours
 		{
 			if(floattime < 10)
-				str_format(stringvar, sizeof(stringvar), "%0d:%0d:0%f", (int)(Time / 60) / 60, (int)Time / 60, floattime);
+				str_format(stringvar, sizeof(stringvar), "%0d:%0d:0%.6f", (int)(Time / 60) / 60, (int)Time / 60, floattime);
 			else
-				str_format(stringvar, sizeof(stringvar), "%0d:%0d:%f", (int)(Time / 60) / 60, (int)Time / 60, floattime);
+				str_format(stringvar, sizeof(stringvar), "%0d:%0d:%.6f", (int)(Time / 60) / 60, (int)Time / 60, floattime);
 		}
 		else
 		{
 			if(floattime < 10)
-				str_format(stringvar, sizeof(stringvar), "%0d:0%f", (int)Time / 60, floattime);
+				str_format(stringvar, sizeof(stringvar), "%0d:0%.6f", (int)Time / 60, floattime);
 			else
-				str_format(stringvar, sizeof(stringvar), "%0d:%f", (int)Time / 60, floattime);
+				str_format(stringvar, sizeof(stringvar), "%0d:%.6f", (int)Time / 60, floattime);
 		}
 	}
 	else
 	{
-		str_format(stringvar, sizeof(stringvar), "%f", floattime);
+		str_format(stringvar, sizeof(stringvar), "%.6f", floattime);
 	}
 
     return stringvar;
